Move Parse impl of SpecifiedUrl into corresponding files.
